+++ src/backends/webgpu/src/kernels/maxpool_webgpu.ts
@@ -16,7 +16,9 @@
  */
 
 import {Conv2DInfo} from '@tensorflow/tfjs-core/dist/ops/conv_util';
-import {getCoordsDataType} from '../shader_preprocessor';
+
+import {generateGetOutputCoords} from '../shader_util';
+
 import {WebGPUProgram} from './webgpu_program';
 
 export class MaxPoolProgram implements WebGPUProgram {
@@ -56,68 +58,6 @@ export class MaxPoolProgram implements WebGPUProgram {
           this.tileSize[2])
     ];
 
-    const generateStrides =
-        (indicesArr: number[], variableName: string): string[] => {
-          if (Math.max(...indicesArr) > 3) {
-            throw new Error('Cannot generate strides for rank > 4.');
-          }
-
-          const rank = indicesArr.length;
-          const dims = ['x', 'y', 'z', 'w'];
-          const shape = indicesArr.map(d => `${variableName}.${dims[d]}`);
-          const strides = new Array(rank - 1);
-          strides[rank - 2] = shape[rank - 1];
-          for (let i = rank - 3; i >= 0; --i) {
-            strides[i] = `(${strides[i + 1]} * ${shape[i + 1]})`;
-          }
-
-          return strides;
-        };
-
-    const generateGetOutputCoords = (rank: number) => {
-      const dtype = getCoordsDataType(rank);
-      const globalInvocation = ['x', 'y', 'z'];
-      let gatherDimensionsStr = '';
-      for (let i = 0; i < dispatchArrangement.length; i++) {
-        const arr = dispatchArrangement[i];
-
-        if (arr.length === 1) {
-          gatherDimensionsStr +=
-              `uint d${arr[0]} = gl_GlobalInvocationID.${globalInvocation[i]};`;
-        } else {
-          const strides = generateStrides(arr, 'outShape');
-          gatherDimensionsStr += `uint index${i} = 
-            gl_GlobalInvocationID.${globalInvocation[i]};`;
-          for (let j = 0; j < strides.length; j++) {
-            gatherDimensionsStr += `
-              uint d${arr[j]} = index${i} / ${strides[j]};
-            `;
-
-            if (j === strides.length - 1) {
-              gatherDimensionsStr += `
-                uint d${arr[j + 1]} = index${i} - d${arr[j]} * ${strides[j]};
-              `;
-            } else {
-              gatherDimensionsStr += `index${i} -= d${arr[j]} * ${strides[j]};`;
-            }
-          }
-        }
-      }
-
-      const dimensions = [];
-      for (let i = 0; i < rank; i++) {
-        dimensions.push(`d${i}`);
-      }
-
-      return `
-        ${dtype} getOutputCoords() {
-          ${gatherDimensionsStr}
-
-          return ${dtype}(${dimensions.join(',')});
-        }
-      `;
-    };
-
     this.userCode = `
       float getValue(int batch, int xR, int xC, int d) {
         if (xC < 0 || xC >= convDims.x) {
@@ -126,7 +66,7 @@ export class MaxPoolProgram implements WebGPUProgram {
         return x[getFlatIndex(ivec4(batch, xR, xC, d), xShape)];
       }
 
-      ${generateGetOutputCoords(this.outputShape.length)}
+      ${generateGetOutputCoords(dispatchArrangement, this.outputShape.length)}
 
       void main() {
         ivec4 coords = getOutputCoords();
